mhm
okay so i implemented the functions which takes in  an image, the "fire.png" image is in my desktop,along with the source files. when i run the server first, and then i run the client, it returns the bytes of the image, extremely long in weird alphabets and stuff. however, i cant find a way to convert the bytes back to the original image, after several google searchings. therefore now, im going to try to convert the image to a str, and then convert the str back to the img.